refactor: Refactor vue template codebase into more modular way
