+++ packages/create-scaffa/templates/vue/src/modules/dashboard/chunks/StatsCards.vue
@@ -1,0 +1,55 @@
+<template>
+  <div class="grid gap-4 md:grid-cols-2 lg:grid-cols-4">
+    <Card>
+      <CardHeader class="flex flex-row items-center justify-between space-y-0 pb-2">
+        <CardTitle class="text-sm font-medium">Total Products</CardTitle>
+        <Package class="h-4 w-4 text-muted-foreground" />
+      </CardHeader>
+      <CardContent>
+        <div class="text-2xl font-bold">{{ data?.length || 0 }}</div>
+        <p class="text-xs text-muted-foreground">+20.1% from last month</p>
+      </CardContent>
+    </Card>
+
+    <Card>
+      <CardHeader class="flex flex-row items-center justify-between space-y-0 pb-2">
+        <CardTitle class="text-sm font-medium">Active Users</CardTitle>
+        <Users class="h-4 w-4 text-muted-foreground" />
+      </CardHeader>
+      <CardContent>
+        <div class="text-2xl font-bold">+2350</div>
+        <p class="text-xs text-muted-foreground">+180.1% from last month</p>
+      </CardContent>
+    </Card>
+
+    <Card>
+      <CardHeader class="flex flex-row items-center justify-between space-y-0 pb-2">
+        <CardTitle class="text-sm font-medium">Sales</CardTitle>
+        <CreditCard class="h-4 w-4 text-muted-foreground" />
+      </CardHeader>
+      <CardContent>
+        <div class="text-2xl font-bold">+12,234</div>
+        <p class="text-xs text-muted-foreground">+19% from last month</p>
+      </CardContent>
+    </Card>
+
+    <Card>
+      <CardHeader class="flex flex-row items-center justify-between space-y-0 pb-2">
+        <CardTitle class="text-sm font-medium">Active Now</CardTitle>
+        <Activity class="h-4 w-4 text-muted-foreground" />
+      </CardHeader>
+      <CardContent>
+        <div class="text-2xl font-bold">+573</div>
+        <p class="text-xs text-muted-foreground">+201 since last hour</p>
+      </CardContent>
+    </Card>
+  </div>
+</template>
+
+<script setup lang="ts">
+import api from '@/api'
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
+import { Package, Users, CreditCard, Activity } from 'lucide-vue-next'
+
+const { data } = api.products.getProducts.useQuery()
+</script>
